feat(F-207): restructure components folder (#269)
* feat(F-207): restructure components folder

* fix(F-207): move header and footer to components folder

---------
