+++ frontend/src/app/[locale]/(main)/new-training/components/CategoryButton.tsx
