+++ frontend/src/app/[locale]/(main)/onboarding/components/UserRadioComponent.tsx
