+++ frontend/src/app/[locale]/(main)/preparation/components/PreparationChecklist.tsx
