+++ frontend/src/app/[locale]/(main)/training-settings/components/UserPreferences.tsx
