+++ frontend/src/app/[locale]/(main)/training-settings/page.tsx
@@ -1,7 +1,7 @@
 import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
 import { MetadataProps } from '@/interfaces/MetadataProps';
-import TrainingSettings from '@/components/common/TrainingSettings';
+import TrainingSettings from './components/TrainingSettings';
 
 export async function generateMetadata({ params }: MetadataProps): Promise<Metadata> {
   const { locale } = await params;
