+++ frontend/src/app/[locale]/(standalone)/simulation/[id]/components/SimulationMessages.tsx
