+++ frontend/src/app/[locale]/(standalone)/simulation/[id]/components/SimulationRealtimeSuggestions.tsx
